% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.R
\name{approx.bathy}
\alias{approx.bathy}
\title{Approximate Bathymetry}
\usage{
approx.bathy(zMax, lkeArea, numZ)
}
\arguments{
\item{zMax}{numeric max depth in meters}

\item{lkeArea}{numeric lake area in meters squared.}

\item{numZ}{integer number of depth layers}
}
\description{
Approximate bathymetry from max depth and lake area.
}
\examples{
\dontrun{
approx.bathy(zMax = 25, lkeArea = 39400000, numZ = 15)
}
}
