% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.R
\name{approx.hypso}
\alias{approx.hypso}
\title{Approximate Hypsography}
\usage{
approx.hypso(zMax, lkeArea, numZ)
}
\arguments{
\item{zMax}{numeric max depth in meters}

\item{lkeArea}{numeric lake area in meters squared.}

\item{numZ}{integer number of depth layers}
}
\description{
Approximate hypsography from max depth and lake area.
}
\examples{
\dontrun{
approx.hypso(zMax = 25, lkeArea = 39400000, numZ = 15)
}
}
