% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm-utils.R
\name{by_s_m}
\alias{by_s_m}
\title{Service subroutine for determining Mixed Layer Depth for a SPECIFIED ERROR NORM VALUE}
\usage{
by_s_m(thres = thres, z0 = z0, zmax = zmax, z = z, sigma = sigma)
}
\arguments{
\item{thres}{error norm;}

\item{z0}{initial depth: use to omit data above z0}

\item{zmax}{maximum depth: use to omit data below zmax}

\item{z}{input x data array, should be increasing function of index}

\item{sigma}{input y data array}
}
\value{
list(nimax=nimax,by_s_m=ss, cline=cline,smz=smz,sms=sms)
\itemize{
\item nimax: number of segments
\item smz: final z array of segmented data
\item sms: final sigma array of segmented data
\item by_s_m: position of MLD = smz(2); or -99 if something is not right
\item cline: Cline depth is defined as the midpoint of the segment connecting inflection points that has the maximum slope
}
}
\description{
Service subroutine for determining Mixed Layer Depth for a SPECIFIED ERROR NORM VALUE
}
