% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_s_m3.R
\name{by_s_m3}
\alias{by_s_m3}
\title{Service subroutine for determining Mixed Layer Depth for a SPECIFIED NUMBER OF SEGMENTS}
\usage{
by_s_m3(nr, z0, zmax, z, sigma)
}
\arguments{
\item{nr}{fixed number of segments}

\item{z0}{initial depth: use to omit data above z0}

\item{zmax}{maximum depth: use to omit data below zmax}

\item{z}{input x data array, should be increasing function of index}

\item{sigma}{input y data array}
}
\value{
list(eps,smz=smz,sms=sms,by_s_m=ss)
eps: the maximum error over all intervals.
smz: final z array of segmented data
sms: final sigma array of segmented data
by_s_m: position of MLD = smz(2); or -99 if something is not right
}
\description{
Service subroutine for determining Mixed Layer Depth for a SPECIFIED ERROR NORM VALUE
FORTRAN comments:
      real FUNCTION BY_s_m3(n,nimax,thres,z0,z,zmax,sigma,smz,sms)
C     (This is the service subroutine for the case of a SPECIFIC ERROR NORM)
c     (Calls subroutine S_mN)
C       Input:
C          N   -[INTEGER] number of points;
C          NR  -[INTEGER] number of segments (fixed);
C          X   -[REAL(N)] input x data array, should be increasing function of index
C          Y   -[REAL(N)] input y data array
C       Output:
C          NI  -[INTEGER] final array with segment start points
}

