% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{earlyspring}
\alias{earlyspring}
\title{Early Spring Profile}
\format{An object of class \code{data.frame} with 604 rows and 6 columns.}
\source{
Supplied
}
\usage{
earlyspring
}
\description{
\describe{
  \item{depth}{depth}
  \item{temper}{temperature}
  \item{salinity}{salinity}
  \item{oxygen}{oxygen}
  \item{oxygen.sat}{oxygen saturation}
  \item{density}{Density}
  ...
}
}
\keyword{datasets}

