% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.temperature.R
\name{epi.temperature}
\alias{epi.temperature}
\title{Get volumetrically averaged epilimnion temp}
\usage{
epi.temperature(wtr, depths, bthA, bthD)
}
\arguments{
\item{wtr}{a water temperature vector}
}
\details{
Calculates volumetrically weighted average epilimnetic temperature using
the supplied water temperature timeseries. If the lake is not stratified, the bottom 
of the epilimnion is calculated as the full depth of the lake.
}
\seealso{
\code{\link{hyp.temperature}}, \code{\link{whole.lake.temperature}}
}

