% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helper.functions.R
\name{get.datetime}
\alias{get.datetime}
\title{Search for and return the datetime column from a ts data.frame}
\usage{
get.datetime(data, error = FALSE)
}
\description{
Search for and return the datetime column from a ts data.frame
}

