% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getxx.R
\name{getxx}
\alias{getxx}
\title{Linear interpolation of the Input Data Series}
\usage{
getxx(x, y, x0)
}
\arguments{
\item{x0}{[REAL] start point along the x-axis}

\item{X}{[REAL(N)] input x-axis array (should be in "chronological" order)}

\item{Y}{[REAL(N)] input y-axis array}
}
\value{
Interpolation of both x and y
}
\description{
Linear interpolation of the Input Data Series
}

