% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getxxnorm.R
\name{getxxnorm}
\alias{getxxnorm}
\title{Interpolate and normalize a vector of samples.}
\usage{
getxxnorm(x, y, x0)
}
\arguments{
\item{x}{[REAL(N)] input x-axis array (should be in "chronological" order)}

\item{y}{[REAL(N)] input y-axis array}

\item{x0}{[REAL] start point along the x-axis}
}
\value{
Output is a data.frame with: anormx, the normalization value used to make the last x == 1. anormy, the normalization value used to make the last y == 1. xx,yy  vectors of x,y values interpolated to be the same length as x0.
}
\description{
Interpolate and normalize a vector of samples. X and Y will be linearly interpolated to match the length of X0 and will be normalized so that the first point is (0,0) and the last point is (1,1). Vector x are the input x values, must be in ascending order.
}
\examples{
x0 <- c(0,1,2,3,4,5)
getxxnorm(y=t11$depth,x=t11$temper, x0=x0)
}

