% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getxxnorm2.R
\name{getxxnorm2}
\alias{getxxnorm2}
\title{Alternate gettxxnorm}
\usage{
getxxnorm2(x, y, dx, nn)
}
\arguments{
\item{x}{[REAL(N)] input x-axis array (should be in "chronological" order)}

\item{y}{[REAL(N)] input y-axis array}

\item{dx}{[REAL] step value for xx; i.e., dx = xx[i+1]-xx[i] for all i<nn}

\item{nn}{[INTEGER] input, desired dimension of output xx,yy vectors
#@param x0 [REAL] start point along the x-axis; i.e., xx[1] = x0}
}
\value{
Output is a data.frame with: anormx, the normalization value used to make the last x == 1.
anormy, the normalization value used to make the last y == 1.
xx,yy  vectors of x,y values interpolated to be the same length as x0

x <- c(1,2,3,3.5)
y <- c(10.0,20.0,10.0,5.0)
n1 <- 10 #desired dimension of output x,y vectors
dx=(max(x)-min(x))/(n1-1) #i.e. step value for x; i.e., dx = x[i+1]-x[i] for all i<n1

getxxnorm2(x=x,y=y,dx=dx, nn=n1)
}
\description{
Alternate gettxxnorm
}

