% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{large_df}
\alias{large_df}
\title{Large df}
\format{A data frame with 20 rows and 2 variables:
\describe{
  \item{x}{depth}
  \item{y}{var}
  ...
}}
\source{
Generated
}
\usage{
large_df
}
\description{
A larger dataset to simulate longer cast cast
}
\keyword{datasets}

