% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{latefall}
\alias{latefall}
\title{Late Fall Profile}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2281 rows and 5 columns.}
\source{
Supplied
}
\usage{
latefall
}
\description{
\describe{
  \item{depth}{depth}
  \item{temper}{temperature}
  \item{salinity}{salinity}
  \item{oxygen}{oxygen}
  \item{oxygen.sat}{oxygen saturation}
  ...
}
}
\keyword{datasets}
