% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{latesummer}
\alias{latesummer}
\title{Late Summer Profile}
\format{An object of class \code{data.frame} with 881 rows and 6 columns.}
\usage{
latesummer
}
\description{
\describe{
  \item{depth}{depth}
  \item{temper}{temperature}
  \item{salinity}{salinity}
  \item{oxygen}{oxygen}
  \item{oxygen.sat}{oxygen saturation}
  \item{density}{Density}
  ...
}
}
\keyword{datasets}
