% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm-algorithm.R
\name{merge_intervals}
\alias{merge_intervals}
\title{merge_intervals}
\usage{
merge_intervals(i, ni)
}
\arguments{
\item{i}{interval number of the first segment to merge}

\item{ni}{current array with interval start point numbers}
}
\description{
merge interval i and i+1
}
\details{
Intervals to be merged
}
