% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_merge.R
\name{mix_merge}
\alias{mix_merge}
\title{mix_merge (was called zerge)}
\usage{
mix_merge(i, nr, ni)
}
\arguments{
\item{i}{[INTEGER] interval number of the first segment to merge}

\item{nr}{[INTEGER] current maximum of the interval number}

\item{ni}{[INTEGER(NR)] current array with interval start point numbers}
}
\value{
'nn- [INTEGER(NR)] new array with interval start point numbers
}
\description{
merge interval i and i+1
}

