% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm-algorithm.R
\name{s_m_p}
\alias{s_m_p}
\title{Computing linear segments for a specified error norm value.}
\usage{
s_m_p(eps, x, y)
}
\arguments{
\item{eps}{[real] error norm}

\item{x}{[real] input x-axis array, should be an increasing function of index}

\item{y}{[real] input y-axis array}
}
\value{
[integer] array A of indices giving data segments.
A[i] start of interval; A[i+1] end of interval, for any i<length(A)
}
\description{
Segments the data in x,y into the intervals given in the output array A.
The data in each interval can be linearly fitted within an error, given by r2b(), less than eps.
}
\keyword{internal}
