% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_df}
\alias{small_df}
\title{Small df}
\format{A data frame with 4 rows and 2 variables:
\describe{
  \item{x}{depth}
  \item{y}{var}
  ...
}}
\source{
Generated
}
\usage{
small_df
}
\description{
A small dataset to simulate short cast
}
\keyword{datasets}

