% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{t11}
\alias{t11}
\title{97130187.t11}
\format{A data frame with 401 rows and 6 variables:
\describe{
  \item{depth}{depth}
  \item{temper}{temperature}
  \item{salinity}{salinity}
  \item{oxygen}{oxygen}
  \item{oxygen.sat}{oxygen saturation}
  \item{density}{Density}
  ...
}}
\source{
Supplied by R Thomson
}
\usage{
t11
}
\description{
Data supplied with original FORTRAN code
}
\keyword{datasets}

