% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rLakeAnalyzer.R
\docType{package}
\name{tidyhydat}
\alias{tidyhydat}
\alias{tidyhydat-package}
\title{Lake Physics Tools}
\description{
Standardized methods for calculating common important derived 
physical features of lakes including water density based based on
temperature, thermal layers, thermocline depth, lake number, Wedderburn
number, Schmidt stability and others.
}
