% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whole.lake.temperature.R
\name{whole.lake.temperature}
\alias{whole.lake.temperature}
\title{Get volumetrically averaged whole lake temperature}
\usage{
whole.lake.temperature(wtr, depths, bthA, bthD)
}
\arguments{
\item{wtr}{a numeric vector of water temperature in degrees C.}

\item{depths}{a numeric vector corresponding to the depths (in m) of the wtr measurements}

\item{bthA}{a numeric vector of cross sectional areas (m^2) corresponding to bthD depths}

\item{bthD}{a numeric vector of depths (m) which correspond to areal measures in bthA}
}
\details{
Calculates volumetrically weighted average whole lake temperature using
the supplied water temperature timeseries.
}
\seealso{
\code{\link{sim_metrics}}, \code{\link{compare_to_field}}, \code{\link{validate_sim}}
}

