% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtr_layer.R
\name{wtr_layer}
\alias{wtr_layer}
\title{Exploration of lake water column layers}
\usage{
wtr_layer(thres = 0.1, z0 = 2.5, zmax = 150, depth = depth,
  measure = measure, nseg = "unconstrained")
}
\arguments{
\item{thres}{error norm; defaults to 0.1}

\item{z0}{initial depth in metres; defaults to 2.5m}

\item{zmax}{maximum depth in metres: default to 150m}

\item{depth}{depth in metres; should be an increasing vector}

\item{measure}{parameter measured in the water column profile}

\item{optional}{parameter to define the number of segments a priori; defaults to an unconstrained approach whereby the algorithm determines segmentations by minimzing the error norm over each segment}
}
\value{
a dataframe of nseg (number of segments), mld (mix layer depth), maxbd (the midpoint of the segment connecting inflection points that has the maximum slope; thermocline for temperature measures)
}
\description{
Extract water column parameters of a given parameter from a profile using the split-and-merge algorithm.
}
\examples{
wtr_layer(depth=latesummer$depth, measure = latesummer$temper)
wtr_layer(depth=latesummer$depth, measure = latesummer$temper, nseg=3)

}
\references{
Thomson & Fine. 2003. Estimating Mixed Layer Depth from Oceanic Profile Data. Journal of Atmospheric and Oceanic Technology. 20(2), 319-329.
}

