% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtr_segments.R
\name{wtr_segments}
\alias{wtr_segments}
\title{Exploration of lake water column segments}
\usage{
wtr_segments(thres = 0.1, z0 = 0.5, zmax = 150, depth = depth,
  measure = measure, nseg = "unconstrained")
}
\arguments{
\item{thres}{error norm; defaults to 0.1}

\item{z0}{initial depth in metres; defaults to auto whereby z0 is calculate as the first value of the longest ordered portion of the depth vector to minimum of 1.}

\item{zmax}{maximum depth in metres: default to 150m}

\item{depth}{depth in metres; should be an increasing vector}

\item{measure}{parameter measured in the water column profile}

\item{nseg}{optional parameter to define the number of segments a priori; defaults to an unconstrained approach whereby the algorithm determines segmentations by minimzing the error norm over each segment}
}
\value{
a dataframe of nseg (number of segments) and the x and y coordinates of the segments produced by the split and merge approach.
}
\description{
Extract water column segments of a given parameter from a profile using the split-and-merge algorithm.
}
\examples{
wtr_segments(depth=latesummer$depth, measure = latesummer$temper)
wtr_segments(depth=latesummer$depth, measure = latesummer$temper, nseg=4)

}
\references{
Thomson & Fine. 2003. Estimating Mixed Layer Depth from Oceanic Profile Data. Journal of Atmospheric and Oceanic Technology. 20(2), 319-329.
}

