% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_s_m.R
\name{by_s_m}
\alias{by_s_m}
\title{Service subroutine for determining Mixed Layer Depth for a SPECIFIED ERROR NORM VALUE}
\usage{
by_s_m(thres, z0, zmax, z, sigma)
}
\arguments{
\item{thres}{error norm;}

\item{z0}{initial depth: use to omit data above z0}

\item{zmax}{maximum depth: use to omit data below zmax}

\item{z}{input x data array, should be increasing function of index}

\item{sigma}{input y data array}
}
\value{
Buncha stuff
}
\description{
Service subroutine for determining Mixed Layer Depth for a SPECIFIED ERROR NORM VALUE
# c     (This subroutine calls the S_M_P subroutine)
c       Input:
C          N     -[INTEGER] number of points;
C          THRES -[REAL]  error norm;
C          Z0    -[REAL]  initial depth: use to omit data above z0
C          ZMAX  -[REAL]  maximum depth: use to omit data below zmax
C
C          Z     -[REAL(N)] input x data array, should be increasing function of index
C          SIGMA -[REAL(N)] input y data array
C       Output:
C          NIMAX -[INTEGER] final number of segments;
C          SMZ   -[REAL(NIMAX)] final z array of segmented data
C          SMS   -[REAL(NIMAX)] final sigma array of segmented data
C          VY_S_M-[REAL] position of MLD (=SMZ(2)).
C                      return -99 if something is not right.
C
}

