% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_s_m3.R
\name{by_s_m3}
\alias{by_s_m3}
\title{Service subroutine for determining Mixed Layer Depth for a SPECIFIED NUMBER OF SEGMENTS}
\usage{
by_s_m3(nr, z0, zmax, z, sigma)
}
\arguments{
\item{nr}{fixed number of segments}

\item{z0}{initial depth: use to omit data above z0}

\item{zmax}{maximum depth: use to omit data below zmax}

\item{z}{input x data array, should be increasing function of index}

\item{sigma}{input y data array}
}
\value{
list(eps=s_mNresults$eps, cline=cline, by_s_m=ss,smz=smz,sms=sms)
\itemize{
\item eps: the maximum error over all intervals.
\item smz: final z array of segmented data
\item sms: final sigma array of segmented data
\item by_s_m: position of MLD = smz(2); or -99 if something is not right
\item cline: Cline depth is defined as the midpoint of the segment connecting inflection points that has the maximum slope
}
}
\description{
Service subroutine for determining Mixed Layer Depth for a SPECIFIED ERROR NORM VALUE
}
\note{
ForTran description - Input (Calls subroutine S_mN)
\itemize{
\item{N -[INTEGER] number of points}
\item{NR  -[INTEGER] number of segments (fixed)}
\item{X   -[REAL(N)] input x data array, should be increasing function of index}
\item{Y   -[REAL(N)] input y data array}
}

ForTran description - Output
\itemize{
\item {NI  -[INTEGER] final array with segment start points}
}
}

