% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.R
\name{get.datetime}
\alias{get.datetime}
\title{Search for and return the datetime column from a ts data.frame}
\usage{
get.datetime(data, error = FALSE)
}
\arguments{
\item{data}{data arg}

\item{error}{defaults to FALSE}
}
\description{
Warns if unavailable then returns NULL.
}
