% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{latesummer}
\alias{latesummer}
\title{Late Summer Profile}
\format{An object of class \code{data.frame} with 881 rows and 6 columns.}
\usage{
latesummer
}
\description{
Late summer water profile taken from Quesnel Lake, British Columbia, Canada. Profile taken with Sea-Bird
  SBE19plus.
}
\details{
\describe{
  \item{depth}{Depth, m}
  \item{temper}{Temperature, degC}
  \item{salinity}{Salinity, PSU}
  \item{oxygen}{Oxygen, ml/l}
  \item{oxygen.sat}{Oxygen saturation, percent saturation}
  \item{density}{Density, kg/m^3}
  ...
}
}
\keyword{datasets}
