% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zerge.R
\name{merge_intervals}
\alias{merge_intervals}
\title{merge_intervals }
\usage{
merge_intervals(i, ni)
}
\arguments{
\item{i}{[INTEGER] interval number of the first segment to merge}

\item{ni}{[INTEGER(NR)] current array with interval start point numbers}
}
\description{
merge intervals i and i+1
}

