% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_m_p.R
\name{s_m_p}
\alias{s_m_p}
\title{Computing linear segments for a specified error norm value.}
\usage{
s_m_p(eps, x, y)
}
\arguments{
\item{eps}{[real] error norm}

\item{x}{[real] input x-axis array, should be an increasing function of index}

\item{y}{[real] input y-axis array}
}
\value{
[integer] array A of indices giving data segments.
A[i] start of interval; A[i+1] end of interval, for any i<length(A)
}
\description{
Segments the data in x,y into the intervals given in the output array A.
The data in each interval can be linearly fitted within an error, given by r2b(), less than eps.
Comments from original Fortran code.
SUBROUTINE s_m_p(n,eps,x,y,Nr,Ni)
C     Main subroutine for determining LINEAR SEGMENTS for a SPECIFIED ERROR NORM VALUE
c     (This subroutine determines the linear fit to the data for a specified error norm)
C     Input:
C          N   -[INTEGER] number of data points;
C          EPS -[REAL]    error norm;
C          X   -[REAL(N)] input x-axis data array, should be increasing function of index
C          Y   -[REAL(N)] input y-axis data array
C       Output:
C          NR  -[INTEGER] final number of segments;
C          NI  -[INTEGER] final array with segment start points
C
INTEGER NR, NI(n)
REAL X(n),Y(n)
}
