% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_mN.R
\name{s_mN}
\alias{s_mN}
\title{Computing linear segments for a specified error norm value.}
\usage{
s_mN(nr, x, y)
}
\arguments{
\item{nr}{[integer] number of segments, fixed.}

\item{x}{[real] input x-axis array, should be an increasing function of index}

\item{y}{[real] input y-axis array}
}
\value{
[list(eps=eps,ni=ni)] where:
eps [real] is the maximum error over all intervals,
ni is a [vector of integer, length nr+1] vector Ni of indices giving data segments
Ni[i] start of interval; Ni[i+1] end of interval, for any i<length(Ni)
}
\description{
Subroutine to determine the Linear SEGMENTS for a PREDEFINED NUMBER OF SEGMENTS (NR)
(Use this program when you want to predefine the number of segments you want to fit
to the data)
Segments the data in x,y into the intervals given in the output array A.
The data in each interval can be linearly fitted within an error, given by r2b(), less than eps.
Comments from original Fortran code:
SUBROUTINE s_mN(n,eps,x,y,Nr,Ni)
C       Input:
C          N   -[INTEGER] number of points;
C          NR  -[INTEGER] number of segments (fixed);
C          X   -[REAL(N)] input x data array, should be increasing function of index
C          Y   -[REAL(N)] input y data array
C       Output:
C        EPS -[REAL] maximum error norm
C          NI  -[INTEGER] final array with segment start points
INTEGER NR, NI(n)
REAL X(n),Y(n)
}

