% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spl.R
\name{spl}
\alias{spl}
\title{spliting interval i into 2 pieces}
\usage{
spl(ni, i)
}
\arguments{
\item{ni}{[INTEGER(NR)] current array with interval start point number}

\item{i}{[INTEGER] interval number, should be less than NR+1}
}
\value{
new array with interval start point number
}
\description{
spliting interval i into 2 pieces
}
\examples{
ni = c(1,5,10,15,20)
nr = 4
ni
ni = spl(ni,i=2)
ni
ni = spl(ni,i=5)
ni
ni = spl(ni,i=2)
ni
ni = spl(ni,i=1)
ni
ni = spl(ni,i=length(ni)-1)
ni
ni = spl(ni,i=length(ni)-1)
ni
ni = spl(ni,i=length(ni)-1)
ni
}

