% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whole.lake.temperature.R
\name{whole.lake.temperature}
\alias{whole.lake.temperature}
\title{Get volumetrically averaged whole lake temperature}
\usage{
whole.lake.temperature(wtr, depths, bthA, bthD)
}
\arguments{
\item{wtr}{a water temperature vector}
}
\details{
A metric function used like an rLakeAnalyzer function to be 
called by \code{\link{sim_metrics}}
}
\seealso{
\code{\link{sim_metrics}}, \code{\link{compare_to_field}}, \code{\link{validate_sim}}
}

